Move ActionLink to shared components
